fix: 修复h3c和华为设备相关的拦截问题
