--- pkg/proxy/interface.go
@@ -1,17 +1,0 @@
-package proxy
-
-import (
-	"github.com/jumpserver/koko/pkg/exchange"
-)
-
-type ParseEngine interface {
-	ParseStream(userInChan chan *exchange.RoomMessage, srvInChan <-chan []byte) (userOut, srvOut <-chan []byte)
-
-	Close()
-
-	NeedRecord() bool
-
-	CommandRecordChan() chan *ExecutedCommand
-
-	RegisterEventCallback(event string, f func())
-}
