+++ pkg/zmodem/events.go
@@ -1,0 +1,9 @@
+package zmodem
+
+type StatusEvent string
+
+const (
+	StartEvent StatusEvent = "ZMODEM_START"
+
+	EndEvent StatusEvent = "ZMODEM_END"
+)
