+++ pkg/zmodem/zmodem_test.go
@@ -1,4 +1,4 @@
-package proxy
+package zmodem
 
 import (
 	"testing"
